refactor flash methods in order to have a helper to share same functions between update and flash
