+++ src/cmd/flash.js
@@ -4,12 +4,12 @@ const ParticleApi = require('./api');
 const VError = require('verror');
 const { HalModuleParser: ModuleParser, ModuleInfo } = require('binary-version-reader');
 const deviceSpecs = require('../lib/device-specs');
-const { ensureError, delay } = require('../lib/utilities');
+const { ensureError } = require('../lib/utilities');
 const { errors: { usageError } } = require('../app/command-processor');
 const dfu = require('../lib/dfu');
 const usbUtils = require('./usb-util');
 const CLICommandBase = require('./base');
-const { platformForId, PLATFORMS } = require('../lib/platform');
+const { platformForId } = require('../lib/platform');
 const settings = require('../../settings');
 const path = require('path');
 const utilities = require('../lib/utilities');
@@ -20,9 +20,7 @@ const { knownAppNames, knownAppsForPlatform } = require('../lib/known-apps');
 const { sourcePatterns, binaryPatterns, binaryExtensions } = require('../lib/file-types');
 const deviceOsUtils = require('../lib/device-os-version-util');
 const semver = require('semver');
-const { moduleTypeToString, sortBinariesByDependency } = require('../lib/dependency-walker');
-
-const FLASH_APPLY_DELAY = 3000;
+const { createFlashSteps, filterModulesToFlash, parseModulesToFlash, flashFiles } = require('../lib/flash-helper');
 
 module.exports = class FlashCommand extends CLICommandBase {
 	constructor(...args) {
@@ -159,7 +157,7 @@ module.exports = class FlashCommand extends CLICommandBase {
 
 		filesToFlash = await this._processBundle({ filesToFlash });
 
-		const fileModules = await this._parseModules({ files: filesToFlash });
+		const fileModules = await parseModulesToFlash({ files: filesToFlash });
 		await this._validateModulesForPlatform({ modules: fileModules, platformId: device.platformId, platformName });
 
 		const deviceOsBinaries = await this._getDeviceOsBinaries({
@@ -170,12 +168,12 @@ module.exports = class FlashCommand extends CLICommandBase {
 			platformId: device.platformId,
 			applicationOnly
 		});
-		const deviceOsModules = await this._parseModules({ files: deviceOsBinaries });
+		const deviceOsModules = await parseModulesToFlash({ files: deviceOsBinaries });
 		let modulesToFlash = [...fileModules, ...deviceOsModules];
-		modulesToFlash = this._filterModulesToFlash({ modules: modulesToFlash, platformId: device.platformId });
+		modulesToFlash = filterModulesToFlash({ modules: modulesToFlash, platformId: device.platformId });
 
-		const flashSteps = await this._createFlashSteps({ modules: modulesToFlash, isInDfuMode: device.isInDfuMode , platformId: device.platformId });
-		await this._flashFiles({ device, flashSteps });
+		const flashSteps = await createFlashSteps({ modules: modulesToFlash, isInDfuMode: device.isInDfuMode , platformId: device.platformId });
+		await flashFiles({ device, flashSteps, ui: this.ui });
 	}
 
 
@@ -405,53 +403,6 @@ module.exports = class FlashCommand extends CLICommandBase {
 		return { module: null, applicationDeviceOsVersion: null };
 	}
 
-	async _flashFiles({ device, flashSteps }) {
-		const progress = this._createFlashProgress({ flashSteps });
-
-		try {
-			for (const step of flashSteps) {
-				if (step.flashMode === 'normal') {
-					if (device.isInDfuMode) {
-						// put device in normal mode
-						progress({ event: 'switch-mode', mode: 'normal' });
-						device = await usbUtils.reopenInNormalMode(device, { reset: true, timeout: 10000 });
-					}
-					// put the device in listening mode to prevent cloud connection
-					try {
-						await device.enterListeningMode();
-					} catch (error) {
-						// ignore if the device is already in listening mode
-					}
-
-
-					// flash the file in normal mode
-					progress({ event: 'flash-file', filename: step.name });
-					await device.updateFirmware(step.data, { progress });
-
-					// wait for the device to apply the firmware
-					await delay(FLASH_APPLY_DELAY);
-					device = await usbUtils.reopenInNormalMode(device, { reset: false });
-				} else {
-					if (!device.isInDfuMode) {
-						// put device in dfu mode
-						progress({ event: 'switch-mode', mode: 'DFU' });
-						device = await usbUtils.reopenInDfuMode(device);
-					}
-
-					// flash the file over DFU
-					progress({ event: 'flash-file', filename: step.name });
-					// CLI always flashes to internal flash which is the DFU alt setting 0
-					const altSetting = 0;
-					await device.writeOverDfu(step.data, { altSetting, startAddr: parseInt(step.moduleInfo.prefixInfo.moduleStartAddy, 16), progress });
-				}
-			}
-		} finally {
-			progress({ event: 'finish' });
-			await device.reset();
-			await device.close();
-		}
-	}
-
 	_createFlashProgress({ flashSteps }) {
 		const NORMAL_MULTIPLIER = 10; // flashing in normal mode is slower so count each byte more
 		const { isInteractive } = this.ui;
@@ -513,69 +464,5 @@ module.exports = class FlashCommand extends CLICommandBase {
 					break;
 			}
 		};
-	}
-
-	async _parseModules({ files }) {
-		return Promise.all(files.map(async (file) => {
-			const parser = new ModuleParser();
-			const binary = await parser.parseFile(file);
-			return {
-				filename: file,
-				...binary
-			};
-		}));
-
-	}
-
-	_filterModulesToFlash({ modules, platformId, allowAll = false }) {
-		const platform = PLATFORMS.find(p => p.id === platformId);
-		const filteredModules = [];
-		// remove encrypted files
-		for (const moduleInfo of modules) {
-			const moduleType = moduleTypeToString(moduleInfo.prefixInfo.moduleFunction);
-			const platformModule = platform.firmwareModules.find(m => m.type === moduleType && m.index === moduleInfo.prefixInfo.moduleIndex);
-			// filter encrypted modules
-			const isEncrypted = platformModule && platformModule.encrypted;
-			const isRadioStack = moduleInfo.prefixInfo.moduleFunction === ModuleInfo.FunctionType.RADIO_STACK;
-			const isNcpFirmware = moduleInfo.prefixInfo.moduleFunction === ModuleInfo.FunctionType.NCP_FIRMWARE;
-			if (!isEncrypted && (!isRadioStack || allowAll) && (!isNcpFirmware || allowAll)) {
-				filteredModules.push(moduleInfo);
-			}
-		}
-		return filteredModules;
-	}
-
-	async _createFlashSteps({ modules, isInDfuMode, platformId }) {
-		const platform = PLATFORMS.find(p => p.id === platformId);
-		const sortedModules = await sortBinariesByDependency(modules);
-		const assetModules = [], normalModules = [], dfuModules = [];
-		sortedModules.forEach(module => {
-			const data = module.prefixInfo.moduleFlags === ModuleInfo.Flags.DROP_MODULE_INFO ? module.fileBuffer.slice(module.prefixInfo.prefixSize) : module.fileBuffer;
-			const flashStep = {
-				name: path.basename(module.filename),
-				moduleInfo: { crc: module.crc, prefixInfo: module.prefixInfo, suffixInfo: module.suffixInfo },
-				data
-			};
-			const moduleType = moduleTypeToString(module.prefixInfo.moduleFunction);
-			const storage = platform.firmwareModules
-				.find(firmwareModule => firmwareModule.type === moduleType);
-			if (moduleType === 'assets') {
-				flashStep.flashMode = 'normal';
-				assetModules.push(flashStep);
-			} else if (moduleType === 'bootloader' || storage.storage === 'external') {
-				flashStep.flashMode = 'normal';
-				normalModules.push(flashStep);
-			} else {
-				flashStep.flashMode = 'dfu';
-				dfuModules.push(flashStep);
-			}
-		});
-
-		// avoid switching to normal mode if device is already in DFU so a device with broken Device OS can get fixed
-		if (isInDfuMode) {
-			return [...dfuModules, ...normalModules, ...assetModules];
-		} else {
-			return [...normalModules, ...dfuModules, ...assetModules];
-		}
 	}
 };
