+++ src/cmd/flash.test.js
@@ -85,33 +85,6 @@ describe('FlashCommand', () => {
 		];
 	};
 
-	const createExtraModules = async () => {
-		const parser = new HalModuleParser();
-		const softDeviceBuffer = await firmwareTestHelper.createFirmwareBinary({
-			moduleFunction: ModuleInfo.FunctionType.RADIO_STACK,
-			moduleIndex: 0,
-			deps: []
-		});
-		const softDevice = await parser.parseBuffer({ fileBuffer: softDeviceBuffer });
-		const ncpBuffer = await firmwareTestHelper.createFirmwareBinary({
-			moduleFunction: ModuleInfo.FunctionType.NCP_FIRMWARE,
-			moduleIndex: 0,
-			deps: []
-		});
-		const ncp = await parser.parseBuffer({ fileBuffer: ncpBuffer });
-		const encryptedModuleBuffer = await firmwareTestHelper.createFirmwareBinary({
-			moduleFunction: ModuleInfo.FunctionType.BOOTLOADER,
-			moduleIndex: 1,
-			deps: []
-		});
-		const encryptedModule = await parser.parseBuffer({ fileBuffer: encryptedModuleBuffer });
-		return {
-			softDevice: { filename: 'softDevice.bin', ...softDevice },
-			ncp: { filename: 'ncp.bin', ...ncp },
-			encryptedModule: { filename: 'encryptedModule.bin', ...encryptedModule }
-		};
-	};
-
 	beforeEach(() => {
 		process.env = {
 			...originalEnv,
@@ -324,23 +297,6 @@ describe('FlashCommand', () => {
 		});
 	});
 
-	describe('_filterModulesToFlash', () => {
-		let modules, assetModules, extraModules;
-		beforeEach( async () => {
-			modules = await createModules();
-			assetModules = await createAssetModules();
-			extraModules = await createExtraModules();
-		});
-		it('returns modules without ncp, softDevice and encrypted modules', async () => {
-			const filteredModules = await flash._filterModulesToFlash({ modules: [...modules, ...assetModules, extraModules.encryptedModule, extraModules.softDevice, extraModules.ncp], platformId: 32 });
-			expect(filteredModules).to.have.lengthOf(7);
-		});
-		it ('returns everything but encrypted modules if allowAll argument is passed', async () => {
-			const filteredModules = await flash._filterModulesToFlash({ modules: [...modules, ...assetModules, extraModules.encryptedModule, extraModules.softDevice, extraModules.ncp], platformId: 32, allowAll: true });
-			expect(filteredModules).to.have.lengthOf(9);
-		});
-	});
-
 	describe('_getDeviceOsBinaries', () => {
 		it('returns empty if there is no application binary', async () => {
 			const modules = await createModules();
@@ -425,140 +381,4 @@ describe('FlashCommand', () => {
 		});
 	});
 
-	describe('_createFlashSteps', () => {
-		let preBootloaderStep, bootloaderStep, systemPart1Step, systemPart2Step, userPart1Step, modules, assetModules, asset1Step, asset2Step;
-		beforeEach(async() => {
-			modules = await createModules();
-			assetModules = await createAssetModules();
-			const preBootloader = modules.find( m => m.filename === 'preBootloader.bin');
-			const bootloader = modules.find( m => m.filename === 'bootloader.bin');
-			const systemPart1 = modules.find( m => m.filename === 'systemPart1.bin');
-			const systemPart2 = modules.find( m => m.filename === 'systemPart2.bin');
-			const userPart1 = modules.find( m => m.filename === 'userPart1.bin');
-			const asset1 = assetModules.find( m => m.filename === 'asset1.bin');
-			const asset2 = assetModules.find( m => m.filename === 'asset2.bin');
-			preBootloaderStep = {
-				name: preBootloader.filename,
-				moduleInfo: {
-					crc: preBootloader.crc,
-					prefixInfo: preBootloader.prefixInfo,
-					suffixInfo: preBootloader.suffixInfo
-				},
-				data: preBootloader.fileBuffer,
-				flashMode: 'normal'
-			};
-			bootloaderStep = {
-				name: bootloader.filename,
-				moduleInfo: {
-					crc: bootloader.crc,
-					prefixInfo: bootloader.prefixInfo,
-					suffixInfo: bootloader.suffixInfo
-				},
-				data: bootloader.fileBuffer,
-				flashMode: 'normal'
-			};
-			systemPart1Step = {
-				name: systemPart1.filename,
-				moduleInfo: {
-					crc: systemPart1.crc,
-					prefixInfo: systemPart1.prefixInfo,
-					suffixInfo: systemPart1.suffixInfo
-				},
-				data: systemPart1.fileBuffer,
-				flashMode: 'dfu'
-			};
-			systemPart2Step = {
-				name: systemPart2.filename,
-				moduleInfo: {
-					crc: systemPart2.crc,
-					prefixInfo: systemPart2.prefixInfo,
-					suffixInfo: systemPart2.suffixInfo
-				},
-				data: systemPart2.fileBuffer,
-				flashMode: 'dfu'
-			};
-			userPart1Step = {
-				name: userPart1.filename,
-				moduleInfo: {
-					crc: userPart1.crc,
-					prefixInfo: userPart1.prefixInfo,
-					suffixInfo: userPart1.suffixInfo
-				},
-				data: userPart1.fileBuffer,
-				flashMode: 'dfu'
-			};
-			asset1Step = {
-				name: asset1.filename,
-				moduleInfo: {
-					crc: asset1.crc,
-					prefixInfo: asset1.prefixInfo,
-					suffixInfo: asset1.suffixInfo
-				},
-				data: asset1.fileBuffer,
-				flashMode: 'normal'
-			};
-			asset2Step = {
-				name: asset2.filename,
-				moduleInfo: {
-					crc: asset2.crc,
-					prefixInfo: asset2.prefixInfo,
-					suffixInfo: asset2.suffixInfo
-				},
-				data: asset2.fileBuffer,
-				flashMode: 'normal'
-			};
-		});
-
-		it('returns a list of flash steps', async () => {
-			const steps = await flash._createFlashSteps({
-				modules,
-				platformId: 6,
-				isInDfuMode: false,
-			});
-
-			const expected = [
-				preBootloaderStep,
-				bootloaderStep,
-				systemPart1Step,
-				systemPart2Step,
-				userPart1Step,
-			];
-			expect(steps).to.deep.equal(expected);
-		});
-
-		it('returns first dfu steps if isInDfuMode is true', async () => {
-			const steps = await flash._createFlashSteps({
-				modules,
-				platformId: 6,
-				isInDfuMode: true,
-			});
-
-			const expected = [
-				systemPart1Step,
-				systemPart2Step,
-				userPart1Step,
-				preBootloaderStep,
-				bootloaderStep,
-			];
-			expect(steps).to.deep.equal(expected);
-		});
-
-		it('returns assets at the end of the list', async () => {
-			const steps = await flash._createFlashSteps({
-				modules: [...assetModules, ...modules],
-				platformId: 6,
-				isInDfuMode: false,
-			});
-			const expected = [
-				preBootloaderStep,
-				bootloaderStep,
-				systemPart1Step,
-				systemPart2Step,
-				userPart1Step,
-				asset2Step,
-				asset1Step,
-			];
-			expect(steps).to.deep.equal(expected);
-		});
-	});
 });
